refactor(payments): lift markRefundAsRefunded into shared lib/payments namespace
Straight relocation — the function was already gateway-agnostic (options
shape has no ShopeePay-specific fields). The original export name is
claimAndSendRefundEmail, so no alias is needed; Task 10 can import
directly. Re-export shim in lib/shopeepay/markRefundAsRefunded.ts keeps
the existing lib/shopeepay/handleRefundNotify.ts import working.
